+++ src/main/java/de/adorsys/keycloak/config/repository/GroupRepository.java
@@ -39,7 +39,8 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import javax.ws.rs.core.Response;
+
+import jakarta.ws.rs.core.Response;
 
 @Service
 public class GroupRepository {