+++ src/main/java/de/adorsys/keycloak/config/repository/IdentityProviderMapperRepository.java
@@ -32,7 +32,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import javax.ws.rs.core.Response;
+
+import jakarta.ws.rs.core.Response;
 
 @Service
 public class IdentityProviderMapperRepository {