+++ src/main/java/de/adorsys/keycloak/config/repository/IdentityProviderRepository.java
@@ -32,8 +32,9 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
-import javax.ws.rs.NotFoundException;
-import javax.ws.rs.core.Response;
+
+import jakarta.ws.rs.NotFoundException;
+import jakarta.ws.rs.core.Response;
 
 @Service
 public class IdentityProviderRepository {