+++ src/main/java/de/adorsys/keycloak/config/repository/UserProfileRepository.java
@@ -30,7 +30,8 @@ import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 
 import java.util.Optional;
-import javax.ws.rs.core.Response;
+
+import jakarta.ws.rs.core.Response;
 
 @Component
 public class UserProfileRepository {