+++ src/main/java/de/adorsys/keycloak/config/repository/UserRepository.java
@@ -32,7 +32,8 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
-import javax.ws.rs.core.Response;
+
+import jakarta.ws.rs.core.Response;
 
 @Service
 public class UserRepository {