+++ src/main/java/de/adorsys/keycloak/config/resource/ManagementPermissions.java
@@ -23,7 +23,12 @@ package de.adorsys.keycloak.config.resource;
 import org.keycloak.representations.idm.ManagementPermissionReference;
 import org.keycloak.representations.idm.ManagementPermissionRepresentation;
 
-import javax.ws.rs.*;
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.PUT;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
 
 /**
  * Routes which are somehow missing from the official Keycloak client library