+++ src/main/java/de/adorsys/keycloak/config/service/clientauthorization/GroupPermissionResolver.java
@@ -26,7 +26,7 @@ import de.adorsys.keycloak.config.repository.GroupRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.ws.rs.NotFoundException;
+import jakarta.ws.rs.NotFoundException;
 
 public class GroupPermissionResolver implements PermissionResolver {
     private static final Logger logger = LoggerFactory.getLogger(GroupPermissionResolver.class);