+++ src/main/java/de/adorsys/keycloak/config/util/resteasy/CookieClientFilter.java
@@ -23,17 +23,23 @@ package de.adorsys.keycloak.config.util.resteasy;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
-import javax.ws.rs.client.ClientRequestContext;
-import javax.ws.rs.client.ClientRequestFilter;
-import javax.ws.rs.client.ClientResponseContext;
-import javax.ws.rs.client.ClientResponseFilter;
+
+import jakarta.ws.rs.client.ClientRequestContext;
+import jakarta.ws.rs.client.ClientRequestFilter;
+import jakarta.ws.rs.client.ClientResponseContext;
+import jakarta.ws.rs.client.ClientResponseFilter;
 
 // By default, the RESTeasy cookie handling is very limited. There is no access to the underlying httpEngine.
 // The underlying httpEngine uses an old CookieSpec by default which is going to refuse valid cookie today.
 // Instead, build the httpEngine from scratch, we are using a RESTeasy filter to grab a re-attach cookie.
 // Currently, this filter does not valide cookie or is able to remove cookies.
 // A cookie managed is required to handle sticky sessions at cookie base
+
 public class CookieClientFilter implements ClientRequestFilter, ClientResponseFilter {
+
+    /**
+     * Holds the additional cookies across multiple interactions in the same thread.
+     */
     private final ThreadLocal<Map<String, String>> cookies = ThreadLocal.withInitial(HashMap::new);
 
     @Override