+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+- Support for Keycloak 22
+
+### Changed
+- Migrated from Java EE to Jakarta EE
+- Migrated imports of javax packages to jakarta packages
+- Upgraded Spring Boot to 2.7.13
+
 ## [5.7.0] - 2023-07-14
 
 ### Changed
@@ -690,8 +698,7 @@ A lot of import properties are added over the years. this major release of keycl
 
 <!-- @formatter:off -->
 
-[Unreleased]: https://github.com/adorsys/keycloak-config-cli/compare/v5.7.0...HEAD
-[5.7.0]: https://github.com/adorsys/keycloak-config-cli/compare/v5.6.1...v5.7.0
+[Unreleased]: https://github.com/adorsys/keycloak-config-cli/compare/v5.6.1...HEAD
 [5.6.1]: https://github.com/adorsys/keycloak-config-cli/compare/v5.6.0...v5.6.1
 [5.6.0]: https://github.com/adorsys/keycloak-config-cli/compare/v5.5.0...v5.6.0
 [5.5.0]: https://github.com/adorsys/keycloak-config-cli/compare/v5.4.0...v5.5.0